turned on warning check for GNU in CI (#94)
* turned on warning check for GNU in CI

* fixed warning

* fixed warning

* changed workflow name

* fixed warning

* fixed more warnings

* fixed warnings

* upgrade to gfortran-10

+++ src/baciof.f90
@@ -253,7 +253,7 @@ SUBROUTINE BAREAD(LU, IB, NB, KA, A)
   LONG_IB = IB
   LONG_NB = NB
   CALL BAREADL(LU, LONG_IB, LONG_NB, LONG_KA, A)
-  KA = LONG_KA
+  KA = INT(LONG_KA)
 END SUBROUTINE BAREAD
 
 !> This subrouytine is using updated baciol() I/O package to read a
@@ -408,7 +408,7 @@ SUBROUTINE BAWRITE(LU, IB, NB, KA, A)
   LONG_IB = IB
   LONG_NB = NB
   CALL BAWRITEL(LU, LONG_IB, LONG_NB, LONG_KA, A)
-  KA = LONG_KA
+  KA = INT(LONG_KA)
 END SUBROUTINE BAWRITE
 
 !> This subrouytine writes a given number of bytes to an unblocked
